# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/mother-cultures": {
      "get": {
        "operationId": "MotherCultureController_getAllMotherCultures",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get all cultures",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MotherCultureListResponseDto"
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Mother Cultures"]
      },
      "post": {
        "operationId": "MotherCultureController_createMotherCulture",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMotherCultureRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create culture",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MotherCultureDetailResponseDto"
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Mother Cultures"]
      }
    },
    "/api/mother-cultures/{id}": {
      "get": {
        "operationId": "MotherCultureController_getMotherCultureById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          }
        ],
        "responses": {
          "200": {
            "description": "Get culture by ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MotherCultureDetailResponseDto"
                }
              }
            }
          },
          "401": { "description": "Unauthorized - you need to be signed in" },
          "404": { "description": "Culture not found" }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Mother Cultures"]
      },
      "put": {
        "operationId": "MotherCultureController_updateMotherCultureById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMotherCultureRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update culture",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MotherCultureDetailResponseDto"
                }
              }
            }
          },
          "400": { "description": "Invalid input data" },
          "404": { "description": "Culture not found" }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Mother Cultures"]
      },
      "delete": {
        "operationId": "MotherCultureController_deleteMotherCultureById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          }
        ],
        "responses": {
          "204": { "description": "Delete culture" },
          "404": { "description": "Culture not found" }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Mother Cultures"]
      }
    },
    "/api/liquid-cultures": {
      "get": {
        "operationId": "LiquidCultureController_getAllLiquidCultures",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get all cultures",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LiquidCultureListResponseDto"
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Liquid Cultures"]
      },
      "post": {
        "operationId": "LiquidCultureController_createLiquidCulture",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLiquidCultureRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create culture",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LiquidCultureDetailResponseDto"
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Liquid Cultures"]
      }
    },
    "/api/liquid-cultures/{id}": {
      "get": {
        "operationId": "LiquidCultureController_getLiquidCultureById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          }
        ],
        "responses": {
          "200": {
            "description": "Get culture by ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LiquidCultureDetailResponseDto"
                }
              }
            }
          },
          "401": { "description": "Unauthorized - you need to be signed in" },
          "404": { "description": "Culture not found" }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Liquid Cultures"]
      },
      "put": {
        "operationId": "LiquidCultureController_updateLiquidCultureById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLiquidCultureRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update culture",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LiquidCultureDetailResponseDto"
                }
              }
            }
          },
          "400": { "description": "Invalid input data" },
          "404": { "description": "Culture not found" }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Liquid Cultures"]
      },
      "delete": {
        "operationId": "LiquidCultureController_deleteLiquidCulture",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          }
        ],
        "responses": {
          "204": { "description": "Delete culture" },
          "404": { "description": "Culture not found" }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Liquid Cultures"]
      }
    },
    "/api/grain-spawns": {
      "get": {
        "operationId": "GrainSpawnController_getAllGrainSpawns",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get all grain spawns",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrainSpawnsListResponseDto"
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Grain Spawns"]
      },
      "post": {
        "operationId": "GrainSpawnController_createGrainSpawn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGrainSpawnRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create grain spawn",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrainSpawnDetailResponseDto"
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Grain Spawns"]
      }
    },
    "/api/grain-spawns/{id}": {
      "get": {
        "operationId": "GrainSpawnController_getGrainSpawnById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          }
        ],
        "responses": {
          "200": {
            "description": "Get grain spawn by ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrainSpawnDetailResponseDto"
                }
              }
            }
          },
          "401": { "description": "Unauthorized - you need to be signed in" },
          "404": { "description": "Grain sapwn not found" }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Grain Spawns"]
      },
      "put": {
        "operationId": "GrainSpawnController_updateGrainSpawnById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGrainSpawnRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update grain spawn",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrainSpawnDetailResponseDto"
                }
              }
            }
          },
          "400": { "description": "Invalid input data" },
          "404": { "description": "Grain spawn not found" }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Grain Spawns"]
      },
      "delete": {
        "operationId": "GrainSpawnController_deleteGrainSpawn",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          }
        ],
        "responses": {
          "204": { "description": "Delete grain spawn" },
          "404": { "description": "Grain spawn not found" }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Grain Spawns"]
      }
    },
...
  },
  "info": {
    "title": "Cultivation App",
    "description": "Cultivation API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" }
    },
    "schemas": {
...
      "SpeciesResponseDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "maxLength": 255 },
          "id": {
            "type": "number",
            "example": 1,
            "description": "Unique identifier of the species"
          }
        },
        "required": ["name", "id"]
      },
...
      "LiquidCultureResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "Unique identifier of the culture"
          },
          "name": {
            "type": "string",
            "example": "Mushroom mania",
            "description": "Name of the culture"
          },
          "inoculationDate": {
            "type": "string",
            "example": "2025-01-20T00:00:00.000Z",
            "description": "Date when culture was inoculated",
            "format": "date-time"
          },
          "contaminationStatus": {
            "type": "boolean",
            "example": true,
            "description": "Contamination status of the culture",
            "nullable": true
          },
          "speciesId": {
            "type": "number",
            "example": 1,
            "description": "Unique identifier of the species"
          }
        },
        "required": [
          "id",
          "name",
          "inoculationDate",
          "contaminationStatus",
          "speciesId"
        ]
      },
      "MotherCultureListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/LiquidCultureResponseDto" }
          }
        },
        "required": ["items"]
      },
      "MotherCultureDetailResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "Unique identifier of the culture"
          },
          "name": {
            "type": "string",
            "example": "Mushroom mania",
            "description": "Name of the culture"
          },
          "inoculationDate": {
            "type": "string",
            "example": "2025-01-20T00:00:00.000Z",
            "description": "Date when culture was inoculated",
            "format": "date-time"
          },
          "speciesId": {
            "type": "number",
            "example": 1,
            "description": "Unique identifier of the species"
          },
          "characteristic": {
            "type": "object",
            "example": "Grow room clone",
            "description": "Extra information about the culture",
            "nullable": true
          },
          "species": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SpeciesResponseDto" }
          }
        },
        "required": [
          "id",
          "name",
          "inoculationDate",
          "speciesId",
          "characteristic",
          "species"
        ]
      },
      "CreateMotherCultureRequestDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "maxLength": 255 },
          "characteristic": { "type": "string", "maxLength": 255 },
          "speciesId": { "type": "number", "minimum": 1 }
        },
        "required": ["name", "characteristic", "speciesId"]
      },
      "UpdateMotherCultureRequestDto": { "type": "object", "properties": {} },
      "LiquidCultureListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/LiquidCultureResponseDto" }
          }
        },
        "required": ["items"]
      },
      "LiquidCultureDetailResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "Unique identifier of the culture"
          },
          "name": {
            "type": "string",
            "example": "Mushroom mania",
            "description": "Name of the culture"
          },
          "inoculationDate": {
            "type": "string",
            "example": "2025-01-20T00:00:00.000Z",
            "description": "Date when culture was inoculated",
            "format": "date-time"
          },
          "contaminationStatus": {
            "type": "boolean",
            "example": true,
            "description": "Contamination status of the culture",
            "nullable": true
          },
          "speciesId": {
            "type": "number",
            "example": 1,
            "description": "Unique identifier of the species"
          },
          "characteristic": {
            "type": "object",
            "example": "USA",
            "description": "Extra information about the culture",
            "nullable": true
          },
          "species": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SpeciesResponseDto" }
          }
        },
        "required": [
          "id",
          "name",
          "inoculationDate",
          "contaminationStatus",
          "speciesId",
          "characteristic",
          "species"
        ]
      },
      "CreateLiquidCultureRequestDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "maxLength": 255 },
          "characteristic": { "type": "string", "maxLength": 255 },
          "contamination status": { "type": "boolean" },
          "speciesId": { "type": "number", "minimum": 1 }
        },
        "required": [
          "name",
          "characteristic",
          "contamination status",
          "speciesId"
        ]
      },
      "UpdateLiquidCultureRequestDto": { "type": "object", "properties": {} },
      "GrainSpawnResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "Unique identifier of the grain spawn"
          },
          "inoculationDate": {
            "type": "string",
            "example": "2025-01-25T00:00:00.000Z",
            "description": "Date when grain spawn was inoculated",
            "format": "date-time"
          },
          "contaminationStatus": {
            "type": "boolean",
            "example": false,
            "description": "Contamination status of the grain spawn",
            "nullable": true
          },
          "shaken": {
            "type": "boolean",
            "example": true,
            "description": "Whether the grain spawn has been shaken",
            "nullable": true
          },
          "motherCultureId": {
            "type": "object",
            "example": 1,
            "description": "ID of the mother culture used",
            "nullable": true
          },
          "liquidCultureId": {
            "type": "object",
            "example": null,
            "description": "ID of the liquid culture used",
            "nullable": true
          },
          "species": { "$ref": "#/components/schemas/SpeciesResponseDto" }
        },
        "required": [
          "id",
          "inoculationDate",
          "contaminationStatus",
          "shaken",
          "motherCultureId",
          "liquidCultureId",
          "species"
        ]
      },
      "GrainSpawnsListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/GrainSpawnResponseDto" }
          }
        },
        "required": ["items"]
      },
      "GrainSpawnDetailResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "Unique identifier of the grain spawn"
          },
          "inoculationDate": {
            "type": "string",
            "example": "2025-01-25T00:00:00.000Z",
            "description": "Date when grain spawn was inoculated",
            "format": "date-time"
          },
          "contaminationStatus": {
            "type": "boolean",
            "example": false,
            "description": "Contamination status of the grain spawn",
            "nullable": true
          },
          "shaken": {
            "type": "boolean",
            "example": true,
            "description": "Whether the grain spawn has been shaken",
            "nullable": true
          },
          "motherCultureId": {
            "type": "object",
            "example": 1,
            "description": "ID of the mother culture used",
            "nullable": true
          },
          "liquidCultureId": {
            "type": "object",
            "example": null,
            "description": "ID of the liquid culture used",
            "nullable": true
          },
          "species": { "$ref": "#/components/schemas/SpeciesResponseDto" },
          "characteristic": {
            "type": "object",
            "example": "This grain has been soaked",
            "description": "Extra characteristics of the grain spawn",
            "nullable": true
          }
        },
        "required": [
          "id",
          "inoculationDate",
          "contaminationStatus",
          "shaken",
          "motherCultureId",
          "liquidCultureId",
          "species",
          "characteristic"
        ]
      },
      "CreateGrainSpawnRequestDto": {
        "type": "object",
        "properties": {
          "speciesId": { "type": "number", "minimum": 1 },
          "characteristic": { "type": "string", "maxLength": 255 },
          "contaminationStatus": { "type": "boolean" },
          "shaken": { "type": "boolean" },
          "motherCultureId": { "type": "number", "minimum": 1 },
          "liquidCultureId": { "type": "number", "minimum": 1 }
        },
        "required": [
          "speciesId",
          "characteristic",
          "contaminationStatus",
          "shaken",
          "motherCultureId",
          "liquidCultureId"
        ]
      },
      "UpdateGrainSpawnRequestDto": { "type": "object", "properties": {} },
...
    }
  }
}